2021.01.10 适配2.0版云盘显示额外信息

+++ jarvins-backend/src/main/java/com/jarvins/entity/file/FileInfo.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -22,7 +23,7 @@ public class FileInfo implements Serializable {
 
     private boolean folder;
 
-    private float size;
+    private BigDecimal size;
 
     private String path;
 
@@ -32,7 +33,7 @@ public class FileInfo implements Serializable {
 
     private LocalDateTime createTime;
 
-    public FileInfo(String name, String fileName, String type, boolean folder, float size, String path, String filePath, String parentPath) {
+    public FileInfo(String name, String fileName, String type, boolean folder, BigDecimal size, String path, String filePath, String parentPath) {
         this.name = name;
         this.fileName = fileName;
         this.type = type;
